Add plugin system
Right now it's hardcoded to look in ~/.shibboleth/plugins, but that
could be configured later. This allows you to extend shibboleth to
do basically whatever you want it to do, without running into much
interference. Shibboleth expects a `handle` function in the plugin
that should take a self (the Shibboleth Cmd loop), and the line.

The function will be attached as `do_<filename>`.